requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.9.1
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cryptography==45.0.6
cycler==0.11.0
datedelta==1.4
dialogflow==2.0.0
flask==3.0.3
fonttools==4.42.0
frozenlist==1.5.0
google-api-core==1.34.1
google-auth==2.40.3
google-cloud-dialogflow==2.41.2
googleapis-common-protos==1.70.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.0.0
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.7.2
multidict==6.1.0
ndg-httpsclient==0.5.1
nltk==3.8.1
numpy==1.24.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
propcache==0.2.0
proto-plus==1.26.1
protobuf==3.20.3
pusher==3.3.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
pyopenssl==25.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2020.1
regex==2023.6.3
requests==2.32.4
rsa==4.9.1
six==1.15.0
textblob==0.17.1
tqdm==4.65.0
twilio==9.7.0
typing-extensions==4.13.2
tzdata==2023.3
unicorn==2.1.3
urllib3==2.2.3
werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
